Dashboard: delete workflow models
The goal of this commit is to delete the workflow-related models in the `main`
Django app used in the dashboard. It is happening from a new migration called
`0032_delete_workflow_models'. The full list is:

- MicroServiceChain
- MicroServiceChainChoice
- MicroServiceChainLink
- MicroServiceChainLinkExitCode
- MicroServiceChoiceReplacementDic
- StandardTaskConfig
- TaskConfig
- TaskConfigAssignMagicLink
- TaskConfigSetUnitVariable
- TaskConfigUnitVariableLinkPull
- TaskType
- WatchedDirectory
- WatchedDirectoryExpectedType

And here is the list of `ForeignKey` fields that had to be migrated to
`UUIDField` as a result of the previous change:

- `job.microservicechainlink`
- `sip.magiclink`
- `transfer.magiclink`
- `unitvariable.microservicechainlink`

Additionally, `0002_initial_data` had to be refactored to load the initial data
using the serialized models instead of using `loaddata`. This was a necessary
change because `loaddata` was only using the current models that were not
present anymore. The fixture `initial_data.json` previously used has been
deleted.